src/main/java/com/gbr/nyan/web/support/Users.java
package com.gbr.nyan.web.support;

import com.gbr.nyan.domain.User;
import org.springframework.security.core.Authentication;
import org.springframework.security.core.context.SecurityContextHolder;

public class Users {

    public static User currentUser() {
        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
        if (authentication == null || !(authentication.getPrincipal() instanceof User)) {
            return null;
        }
        return (User) authentication.getPrincipal();
    }

    public static boolean userIsLoggedIn() {
        return currentUser() != null;
    }
}
